rename faculty folder name
