+++ design-at-cornell/src/design-community/design-faculty/Faculty.tsx
