+++ design-at-cornell/src/design-community/design-faculty/FacultyBubble.tsx
